:hammer: Code refactoring Add unit test
