+++ src/constants.ts
@@ -90,4 +90,4 @@ export const DEFAULT_LISK_CORE_PATH = '~/.lisk/lisk-core';
 export const LEGACY_DB_PATH = `${DEFAULT_LISK_CORE_PATH}/${DEFAULT_DATA_DIR}/legacy.db`;
 
 export const DEFAULT_VERSION = '0.1.0';
-export const NEW_BLOCK_EVENT_NAME = 'app:block:new';
+export const EVENT_NEW_BLOCK = 'app:block:new';
